Reorganize api into auth and sandbox
